Solve issues related DataParallel Training
The original code base had plenty problems with distributed training.
The main code base was written using DistributedDataParallel. This was
not compatible with the machine I was using (problemsLog). I changed the
code base so that it works with DataParallel. Used custom SyncBatchNorm
(picked from the DeepLab codebase). Still not error proof, but the train
sript works.